+++ src/dipdup/indexes/evm_subsquid_events/index.py
@@ -21,6 +21,7 @@ from dipdup.models.evm_subsquid import SubsquidEvent
 from dipdup.models.evm_subsquid import SubsquidEventData
 from dipdup.models.evm_subsquid import SubsquidMessageType
 from dipdup.performance import metrics
+from dipdup.prometheus import Metrics
 
 LEVEL_BATCH_TIMEOUT = 1
 NODE_SYNC_LIMIT = 128
@@ -100,6 +101,8 @@ class SubsquidEventsIndex(
 
         for message_level, level_logs in logs_by_level.items():
             await self._process_level_events(tuple(level_logs), self.topics, message_level)
+            if self._config.expose_metrics:
+                Metrics.set_sqd_processor_last_block(message_level)
 
     def get_sync_level(self) -> int:
         """Get level index needs to be synchronized to depending on its subscription status"""
@@ -131,6 +134,8 @@ class SubsquidEventsIndex(
             return
 
         subsquid_sync_level = await self.datasource.get_head_level()
+        if self._config.expose_metrics:
+            Metrics.set_sqd_processor_chain_height(subsquid_sync_level)
 
         use_node = False
         if self.node_datasources:
@@ -167,6 +172,8 @@ class SubsquidEventsIndex(
                 timestamp = int(block['timestamp'], 16)
                 parsed_level_logs = tuple(EvmNodeLogData.from_json(log, timestamp) for log in level_logs)
                 await self._process_level_events(parsed_level_logs, self.topics, sync_level)
+                if self._config.expose_metrics:
+                    Metrics.set_sqd_processor_last_block(level)
 
         else:
             sync_level = min(sync_level, subsquid_sync_level)
@@ -174,6 +181,8 @@ class SubsquidEventsIndex(
 
             async for _level, events in fetcher.fetch_by_level():
                 await self._process_level_events(events, self.topics, sync_level)
+                if self._config.expose_metrics:
+                    Metrics.set_sqd_processor_last_block(_level)
 
         await self._exit_sync_state(sync_level)