+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## Unreleased
+
+### Added
+
+- evm.subsquid.events: Added metrics for subsquid cloud deploy
+
 ## [7.1.0] - 2023-10-27
 
 ### Added